chore: rename zodConfig.ts -> configureZod.ts
